+++ netbox_ssl/api/serializers/certificates.py
@@ -27,6 +27,8 @@ class CertificateSerializer(NetBoxModelSerializer):
     assignment_count = serializers.SerializerMethodField()
     chain_is_valid = serializers.BooleanField(read_only=True)
     chain_needs_validation = serializers.BooleanField(read_only=True)
+    acme_renewal_due = serializers.BooleanField(read_only=True)
+    acme_renewal_status = serializers.CharField(read_only=True)
 
     class Meta:
         model = Certificate
@@ -62,6 +64,17 @@ class CertificateSerializer(NetBoxModelSerializer):
             "chain_depth",
             "chain_is_valid",
             "chain_needs_validation",
+            # ACME fields
+            "is_acme",
+            "acme_provider",
+            "acme_account_email",
+            "acme_challenge_type",
+            "acme_server_url",
+            "acme_auto_renewal",
+            "acme_last_renewed",
+            "acme_renewal_days",
+            "acme_renewal_due",
+            "acme_renewal_status",
             "tags",
             "custom_fields",
             "created",
@@ -77,6 +90,7 @@ class CertificateSerializer(NetBoxModelSerializer):
             "valid_to",
             "days_remaining",
             "chain_status",
+            "is_acme",
         ]
 
     def get_assignment_count(self, obj):